diagrams/requirements.txt
diagrams
pyyaml
